move client, initial and node interfaces outside main
