+++ UI/mainUI.py
@@ -4,6 +4,8 @@ from qt_interfaces.main_menu_ui import Ui_MainMenu
 from utilities.account_manager import AccountManager
 from engine import StorjEngine
 
+from client_config import ClientConfigurationUI
+
 
 global html_format_begin, html_format_end
 html_format_begin = "<html><head/><body><p><span style=\" font-size:12pt; font-weight:600;\">"
