+++ UI/node_details.py
@@ -1,0 +1,95 @@
+from PyQt4 import QtCore, QtGui
+from qt_interfaces.node_details_ui import Ui_NodeDetails
+from engine import StorjEngine
+from utilities.tools import Tools
+
+from ipwhois import IPWhois
+import pycountry
+import threading
+
+
+global html_format_begin, html_format_end
+html_format_begin = "<html><head/><body><p><span style=\" font-size:12pt; font-weight:600;\">"
+html_format_end = "</span></p></body></html>"
+
+
+# Node details section
+class NodeDetailsUI(QtGui.QMainWindow):
+
+    def __init__(self, parent=None, nodeid=None):
+        QtGui.QWidget.__init__(self, parent)
+
+        self.storj_engine = StorjEngine()  # init StorjEngine
+        # login UI
+        self.node_details_ui = Ui_NodeDetails()
+        self.node_details_ui.setupUi(self)
+
+        self.nodeid = nodeid
+        self.tools = Tools()
+
+        QtCore.QObject.connect(self.node_details_ui.ok_bt, QtCore.SIGNAL("clicked()"), self.close)  # close window
+
+        self.createNewNodeDetailsResolveThread()
+        # print nodeid
+
+    def createNewNodeDetailsResolveThread(self):
+        download_thread = threading.Thread(target=self.initialize_node_details, args=())
+        download_thread.start()
+
+    def initialize_node_details(self):
+        self.node_details_content = self.storj_engine.storj_client.contact_lookup(str(self.nodeid))
+
+        self.node_details_ui.address_label.setText(
+            html_format_begin + str(self.node_details_content.address) + html_format_end)  # get given node address
+        self.node_details_ui.last_timeout_label.setText(
+            html_format_begin + str(self.node_details_content.lastTimeout) + html_format_end)  # get last timeout
+        self.node_details_ui.timeout_rate_label.setText(
+            html_format_begin + str(self.node_details_content.timeoutRate) + html_format_end)  # get timeout rate
+        self.node_details_ui.user_agent_label.setText(
+            html_format_begin + str(self.node_details_content.userAgent) + html_format_end)  # get user agent
+        self.node_details_ui.protocol_version_label.setText(
+            html_format_begin + str(self.node_details_content.protocol) + html_format_end)  # get protocol version
+        self.node_details_ui.response_time_label.setText(html_format_begin + str(
+            self.node_details_content.responseTime) + html_format_end)  # get farmer node response time
+        self.node_details_ui.port_label.setText(
+            html_format_begin + str(self.node_details_content.port) + html_format_end)  # get farmer node port
+        self.node_details_ui.node_id_label.setText(
+            html_format_begin + str(self.nodeid) + html_format_end)  # get farmer node response time
+
+        # ping_to_node = self.tools.measure_ping_latency(str(self.node_details_content.address))
+
+        ip_addr = socket.gethostbyname(str(self.node_details_content.address))
+
+        obj = IPWhois(ip_addr)
+        res = obj.lookup_whois()
+        country = res["nets"][0]['country']
+
+        country_parsed = pycountry.countries.get(alpha_2=str(country))
+
+        country_full_name = country_parsed.name
+
+        self.node_details_ui.country_label.setText(
+            html_format_begin + str(country_full_name) + html_format_end)  # set full country name
+
+        ### Display country flag ###
+
+        self.scene = QtGui.QGraphicsScene()
+
+        # scene.setSceneRect(-600,-600, 600,600)
+        # self.scene.setSceneRect(-600, -600, 1200, 1200)
+
+        # pic = QtGui.QPixmap("PL.png")
+        # self.scene.addItem(QtGui.QGraphicsPixmapItem(pic))
+        # self.view = self.node_details_ui.country_graphicsView
+        # self.view.setScene(self.scene)
+        # self.view.setRenderHint(QtGui.QPainter.Antialiasing)
+        # self.view.show()
+
+        grview = self.node_details_ui.country_graphicsView()
+        scene = QtGui.QGraphicsScene()
+        scene.addPixmap(QPixmap('PL.png'))
+        grview.setScene(scene)
+
+        grview.show()
+
+        print country_full_name
