+++ UI/qt_interfaces/node_details_ui.py
