+++ UI/utilities/backend_config.py
@@ -1,3 +1,6 @@
+import xml.etree.cElementTree as ET
+
+
 # Configuration backend section
 class Configuration():
